# Speed up computation of contact jacobians